theapp/randomapp/requirements.txt
Django==2.0.3
Faker==0.8.12
psycopg2-binary==2.7.4
djangorestframework==3.7.7
